# shared/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: wtrfll API
  description: REST API for sessions, bibles, and passage retrieval.
  version: 0.1.0
  x-contractVersion: 3
servers:
  - url: http://localhost:5000
    description: LAN default (server hosts API + static web)
paths:
  /api/sessions:
    post:
      summary: Create a new session
      operationId: createSession
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/sessions/upcoming:
    get:
      summary: List upcoming sessions
      operationId: listUpcomingSessions
      responses:
        '200':
          description: Upcoming sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpcomingSession'
  /api/sessions/{id}/join:
    post:
      summary: Join an existing session
      operationId: joinSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRequest'
      responses:
        '200':
          description: Joined session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinResult'
        '409':
          description: Controller already locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinResult'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/bibles:
    get:
      summary: List available translations
      operationId: listBibles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bible'
  /api/bible-books:
    get:
      summary: List normalized bible book metadata
      operationId: listBibleBooks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BibleBookMetadata'
  /api/passage:
    get:
      summary: Retrieve a passage by reference and translation
      operationId: getPassage
      parameters:
        - name: ref
          in: query
          required: true
          schema:
            type: string
          description: Scripture reference (e.g., 'Jn 3:16-18')
        - name: translation
          in: query
          required: true
          schema:
            type: string
          description: Translation code (e.g., 'RVR1960', 'NTV')
      responses:
        '200':
          description: Passage result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassageResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/lyrics:
    get:
      summary: List lyrics entries
      operationId: listLyricsEntries
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Optional keyword (title, author, or lyrics text)
      responses:
        '200':
          description: Songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LyricsEntrySummary'
    post:
      summary: Create a lyrics entry
      operationId: createLyricsEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LyricsEntryUpsert'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LyricsEntryDetail'
  /api/lyrics/{id}:
    get:
      summary: Retrieve a lyrics entry
      operationId: getLyricsEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Song
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LyricsEntryDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a lyrics entry
      operationId: updateLyricsEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LyricsEntryUpsert'
      responses:
        '200':
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LyricsEntryDetail'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
    SessionCreated:
      type: object
      properties:
        id:
          type: string
          format: uuid
        shortCode:
          type: string
        controllerJoinToken:
          type: string
        displayJoinToken:
          type: string
        createdAt:
          type: string
          format: date-time
        name:
          type: string
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required: [id, shortCode, controllerJoinToken, displayJoinToken, createdAt, name]
    CreateSessionRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        scheduledAt:
          type: string
          format: date-time
          nullable: true
    JoinRequest:
      type: object
      properties:
        role:
          type: string
          enum: [controller, display]
        joinToken:
          type: string
      required: [role, joinToken]
    JoinResult:
      type: object
      properties:
        ok:
          type: boolean
        controllerLocked:
          type: boolean
        message:
          type: string
        sessionId:
          type: string
          format: uuid
        shortCode:
          type: string
        role:
          type: string
          enum: [controller, display]
        name:
          type: string
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required: [ok, controllerLocked, sessionId, shortCode, role, name]
    UpcomingSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        shortCode:
          type: string
        controllerJoinToken:
          type: string
        displayJoinToken:
          type: string
        createdAt:
          type: string
          format: date-time
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required: [id, name, shortCode, controllerJoinToken, displayJoinToken, createdAt]
    Bible:
      type: object
      properties:
        code:
          type: string
          example: RVR1960
        name:
          type: string
          example: Reina-Valera 1960
        language:
          type: string
          example: es
        provider:
          type: string
          example: external:RVR1960Provider
        cachePolicy:
          type: string
          example: no-store
        license:
          type: string
        version:
          type: string
        isOfflineReady:
          type: boolean
          description: true when the translation is bundled locally and works without network access
      required: [code, name, language, provider]
    BibleBookMetadata:
      type: object
      properties:
        id:
          type: string
        englishDisplayName:
          type: string
        spanishDisplayName:
          type: string
        slug:
          type: string
        ordinal:
          type: integer
          nullable: true
        aliases:
          type: array
          items:
            type: string
      required: [id, englishDisplayName, spanishDisplayName, slug, aliases]
    PassageResult:
      type: object
      properties:
        reference:
          type: string
          example: Juan 3:16-18
        translation:
          type: string
          example: RVR1960
        verses:
          type: array
          items:
            $ref: '#/components/schemas/Verse'
        attribution:
          $ref: '#/components/schemas/Attribution'
        cachePolicy:
          type: string
          example: no-store
      required: [reference, translation, verses]
    Verse:
      type: object
      properties:
        book:
          type: string
          example: Juan
        chapter:
          type: integer
          example: 3
        verse:
          type: integer
          example: 16
        text:
          type: string
      required: [book, chapter, verse, text]
    Attribution:
      type: object
      properties:
        required:
          type: boolean
        text:
          type: string
        url:
          type: string
      required: [required]
    LyricsEntrySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        author:
          type: string
          nullable: true
      required: [id, title]
    LyricsEntryDetail:
      allOf:
        - $ref: '#/components/schemas/LyricsEntrySummary'
        - type: object
          properties:
            lyricsChordPro:
              type: string
          required: [lyricsChordPro]
    LyricsEntryUpsert:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
          nullable: true
        lyricsChordPro:
          type: string
      required: [title, lyricsChordPro]